- Make openDialog return the id of the dialog. This way we can close opened dialogs later again from the outside - Move the processQrDialog to the processOpenQrDialog method
